added necessary project files, added gitignore file for unwanted files
